refactor: split rich_text_editor into submodules
Split the monolithic 2808-line rich_text_editor/mod.rs into focused modules
following single responsibility principle.

## New Modules

### core.rs (359 lines)
- RichTextEditor struct definition with fields marked pub(super)
- Constructor: new()
- Builder methods: content(), from_markdown(), view_mode(), toolbar(), focused(), bg(), fg()
- Getters: cursor_position(), set_cursor(), block_count()
- Link/Image operations: open_link_dialog(), open_image_dialog(), insert_link(), insert_image()
- Dialog operations: is_dialog_open(), close_dialog()
- impl Default, impl_styled_view!, impl_props_builders! macros

### editing.rs (693 lines)
- Cursor movement: move_left/right/up/down, move_home/end, move_document_start/end
- Selection: start_selection(), clear_selection(), has_selection(), get_selection(), delete_selection()
- Text editing: insert_char(), insert_str(), delete_char_before(), delete_char_at()
- Block operations: split_block(), merge_with_previous/next(), delete_block()
- Formatting: toggle_bold/italic/underline/strikethrough/code()
- Block types: set_block_type(), current_block_type(), current_format()
- Undo/redo: undo(), redo()
- Toolbar actions: toolbar_action(), process_markdown_shortcuts()
- Key handling: handle_key(), handle_dialog_key()

### rendering.rs (506 lines)
- render_toolbar() - Formatting toolbar with active state
- render_editor() - Main editor view with cursor and selection
- render_preview() - Preview mode with formatted output
- render_heading() - Heading rendering with modifiers
- render_dialog() - Insert link/image dialog UI

### markdown.rs (129 lines)
- parse_markdown() - Parse full markdown document
- parse_markdown_line() - Parse single line to Block
- Handles: headings, quotes, lists, code blocks, horizontal rules

## mod.rs Changes

- Reduced from 2808 lines to 1293 lines (-54%)
- Added module declarations: mod core, editing, markdown, rendering
- Kept: EditOp and DialogType enums, impl View, tests
- Re-export: pub use core::RichTextEditor;
- Removed extracted functions (now in separate modules)

## Benefits

- **Maintainability**: Each module has a single, clear responsibility
- **Navigation**: Easier to find specific functionality
- **Testing**: Smaller, focused modules are easier to test
- **Code Review**: Changes are scoped to specific modules

Fixes #247
